add/edit/discard buttons wired up
