+++ src/app/home/pages/PlansPage/components/Wizard/NamespacesTable.module.scss
@@ -1,0 +1,3 @@
+.actionsContainer {
+  min-width: 4em;
+}
